+++ src/vojsic_tree/mod.rs
@@ -1,4 +1,4 @@
-pub fn solution(n: usize) -> usize {
+pub fn solution(n: usize) -> u128 {
     let mut res = 0;
 
     if n == 0 {
@@ -7,13 +7,13 @@ pub fn solution(n: usize) -> usize {
 
     let x = std::cmp::min(3, n - 1);
     for k in 0..(x + 1) {
-        res += solution(k)* solution(n  - 1 - k);
+        res += solution(k) * solution(n - 1 - k);
     }
     res
 }
 
-pub fn solution_optimized(n: usize) -> usize {
-    let mut result = vec![1; n + 1];
+pub fn solution_optimized(n: usize) -> u128 {
+    let mut result = vec![1 as u128; n + 1];
 
     for i in 2..(n + 1) {
         result[i] = 0;
